boodle: renommage des questions; boodle boss.php: securisation des init tables (GET a ecrire a la main)

+++ ink/boodladm.php
@@ -3,25 +3,34 @@
 class boodladm extends boodle
 {
 // cree toutes les tables d'un boodle de base
-function create_tables() {
+function create_tables( $tab ) {
 	// table des logins (non editable sous longage)
-	$sqlrequest = "DROP TABLE IF EXISTS `{$this->table_logins}`";
-	$result = $this->db->conn->query( $sqlrequest );
-	if	(!$result) mostra_fatal( $sqlrequest . "<br>" . mysqli_error($this->db->conn) );
-	$sqlrequest = "CREATE TABLE `{$this->table_logins}`
-		( `uchave` VARCHAR(32), `binome` INT,
-		PRIMARY KEY (`uchave`), INDEX(`binome`) )
-		ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_ci;";
-	$result = $this->db->conn->query( $sqlrequest );
-	if	(!$result) mostra_fatal( $sqlrequest . "<br>" . mysqli_error($this->db->conn) );
-	echo "<p>$sqlrequest</p>\n";
+	if	( $tab == 'logins' )
+		{
+		$sqlrequest = "DROP TABLE IF EXISTS `{$this->table_logins}`";
+		$result = $this->db->conn->query( $sqlrequest );
+		if	(!$result) mostra_fatal( $sqlrequest . "<br>" . mysqli_error($this->db->conn) );
+		$sqlrequest = "CREATE TABLE `{$this->table_logins}`
+			( `uchave` VARCHAR(32), `binome` INT,
+			PRIMARY KEY (`uchave`), INDEX(`binome`) )
+			ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_ci;";
+		$result = $this->db->conn->query( $sqlrequest );
+		if	(!$result) mostra_fatal( $sqlrequest . "<br>" . mysqli_error($this->db->conn) );
+		echo "<p>$sqlrequest</p>\n";
+		}
 	// table des binomes
-	global $form_bi;
-	$form_bi->mk_table( $this->db, $this->table_binomes, true );
+	else if	( $tab == 'binomes' )
+		{
+		global $form_bi;
+		$form_bi->mk_table( $this->db, $this->table_binomes, true );
+		}
 	// tables des experiences
-	global $formexp;
-	for	( $i = 1; $i <= 5; $i++ )
-		$formexp[$i]->mk_table( $this->db, $this->table_exp[$i], true, false );
+	else 	{
+		global $formexp;
+		$i = (int)$tab;
+		if	( ( $i >= 1 ) && ( $i <= 5 ) )
+			$formexp[$i]->mk_table( $this->db, $this->table_exp[$i], true, false );
+		}
 	}
 
 // // // objet binome // // //
